refactor: move HOC to hoc folder
