+++ src/components/countryItem/CountryName.tsx
@@ -1,0 +1,14 @@
+/* eslint-disable react/prefer-stateless-function */
+import React from 'react';
+
+class CountryName extends React.Component {
+  render() {
+    return (
+      <div>
+        <input type="text" />
+      </div>
+    );
+  }
+}
+
+export default CountryName;
