+++ src/const.ts
@@ -1,6 +1,7 @@
 export const LINKS = {
   Main: '/',
   'About Us': 'about',
+  Forms: 'forms',
 };
 
 export const SMTH = {};
