+++ src/hoc/withRouterHOC.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
-import { WithRouterProps } from '../../interfaces/interfaces';
+import { WithRouterProps } from '../interfaces/interfaces';
 
 const withRouter = <Props extends WithRouterProps>(
   Component: React.ComponentType<Props>
